Creating button to new page
