+++ TGDP_Website/Run_Website.html
@@ -4,6 +4,21 @@
     <title> Diplomacy </title>
     <meta charset = "UTF-8">
     <meta name = "viewpoint" content = "width=device-width, initial-scale = 1">
+    <style>
+        .button-class {
+            background-color: #4caf50;
+            border: black;
+            color: white;
+            padding: 10px 20px;
+            text-align: center;
+            text-decoration: none;
+            display: inline-block;
+            font-size: 16px;
+            margin: 4px 2px;
+            cursor: pointer;
+            border-radius: 4px;
+        }
+    </style>
 <style>
     body {
         font-family: Arial;
@@ -75,24 +90,19 @@
 
     <button onclick = "getUsers()">Get user data</button>
     <div id = "result-container"></div>
-        <!--
-        onclick = "document.getElementById('get pet owners')">
-        Click me</button>
-    <p id = "demo"></p>
-    -->
-    <!--
-    <div class = "footer">
-        <h2>Huge thank you to Mercy for helping me along the way.</h2>
-    </div>
-    -->
-
-    <!--
-    <img src = "{{image_url}}" alt = "Image from external source">
-    <img src = "{{url_for (filename = 'Europe_Map_For_Website.png')}}" 
-    align = "middle"/>
-    <img src = "{{url_for ('Static', filename = 'Static/Europe_Map_For_Website.png')}}" class = "image" />
-    -->
     <img src = 'Static/Europe_Map_For_Website.png' class = "image" />
+
+
+
+
+
+
+    <!-- TRYING TO MAKE LINK TO ANOTHER PAGE -->
+    <div>
+        <a href = "game8.html" class = "button-class"> </a>a>
+        </a>
+    </div>
+
 
 <script>
